Implemeted review comments. Also some minor code refactoring.
