+++ bigtable/google/cloud/bigtable/row_set.py
@@ -16,8 +16,6 @@
 
 
 from google.cloud._helpers import _to_bytes
-from google.cloud.bigtable_v2.proto import (
-    bigtable_pb2 as data_messages_v2_pb2)
 
 
 class RowSet(object):
@@ -91,85 +89,15 @@ class RowRange(object):
         google.bigtable.v2.RowRange add method.
         """
         range_kwargs = {}
-        start_key_key = 'start_key_open'
-        if self.start_inclusive:
-            start_key_key = 'start_key_closed'
-        range_kwargs[start_key_key] = _to_bytes(self.start_key)
-        end_key_key = 'end_key_open'
-        if self.end_inclusive:
-            end_key_key = 'end_key_closed'
-        range_kwargs[end_key_key] = _to_bytes(self.end_key)
+        if self.start_key is not None:
+            start_key_key = 'start_key_open'
+            if self.start_inclusive:
+                start_key_key = 'start_key_closed'
+            range_kwargs[start_key_key] = _to_bytes(self.start_key)
+
+        if self.end_key is not None:
+            end_key_key = 'end_key_open'
+            if self.end_inclusive:
+                end_key_key = 'end_key_closed'
+            range_kwargs[end_key_key] = _to_bytes(self.end_key)
         return range_kwargs
-
-
-class ReadRowsRequestManager(object):
-    """ Update the ReadRowsRequest message in case of failures by
-        filtering the already read keys.
-
-    :type message: class:`data_messages_v2_pb2.ReadRowsRequest`
-    :param message: Original ReadRowsRequest containing all of the parameters
-                    of API call
-
-    :type last_scanned_key: bytes
-    :param last_scanned_key: last successfully scanned key
-
-    :type rows_read_so_far: int
-    :param rows_read_so_far: total no of rows successfully read so far.
-                            this will be used for updating rows_limit
-
-    """
-
-    def __init__(self, message, last_scanned_key, rows_read_so_far):
-        self.message = message
-        self.last_scanned_key = last_scanned_key
-        self.rows_read_so_far = rows_read_so_far
-        self.new_message = None
-
-    def build_updated_request(self):
-        """ Updates the given message request as per last scanned key
-        """
-        r_kwargs = {'table_name': self.message.table_name}
-        r_kwargs['filter'] = self.message.filter
-
-        if self.message.rows_limit != 0:
-            r_kwargs['rows_limit'] = max(1, self.message.rows_limit -
-                                         self.rows_read_so_far)
-
-        self.new_message = data_messages_v2_pb2.ReadRowsRequest(**r_kwargs)
-        self._filter_rows_keys()
-        self._filter_row_ranges()
-        return self.new_message
-
-    def _filter_rows_keys(self):
-        """ Helper for :meth:`build_updated_request`"""
-        for row_key in self.message.rows.row_keys:
-            if(row_key > self.last_scanned_key):
-                self.new_message.rows.row_keys.append(row_key)
-
-    def _filter_row_ranges(self):
-        """ Helper for :meth:`build_updated_request`"""
-        new_row_ranges = []
-
-        for row_range in self.message.rows.row_ranges:
-            if((row_range.end_key_open and
-                self._key_already_read(row_range.end_key_open)) or
-                (row_range.end_key_closed and
-                 self._key_already_read(row_range.end_key_closed))):
-                    continue
-
-            if ((row_range.start_key_open and
-                self._key_already_read(row_range.start_key_open)) or
-                (row_range.start_key_closed and
-                 self._key_already_read(row_range.start_key_closed))):
-                row_range.start_key_closed = _to_bytes("")
-                row_range.start_key_open = self.last_scanned_key
-
-                new_row_ranges.append(row_range)
-            else:
-                new_row_ranges.append(row_range)
-
-        self.new_message.rows.row_ranges.extend(new_row_ranges)
-
-    def _key_already_read(self, key):
-        """ Helper for :meth:`_filter_row_ranges`"""
-        return key <= self.last_scanned_key
