+++ bigtable/tests/unit/test_row_set.py
@@ -14,7 +14,6 @@
 
 
 import unittest
-import mock
 from google.cloud.bigtable.row_set import RowRange
 from google.cloud._helpers import _to_bytes
 
@@ -102,159 +101,6 @@ class TestRowRange(unittest.TestCase):
         row_range = self._make_one(start_key, end_key, False, True)
         actual_result = row_range.get_range_kwargs()
         self.assertEqual(expected_result, actual_result)
-
-
-class TestReadRowsRequestManager(unittest.TestCase):
-
-    @classmethod
-    def setUpClass(cls):
-        cls.table_name = 'table_name'
-        cls.row_range1 = RowRange("row_key21", "row_key29")
-        cls.row_range2 = RowRange("row_key31", "row_key39")
-        cls.row_range3 = RowRange("row_key41", "row_key49")
-
-        cls.request = _ReadRowsRequestPB(table_name=cls.table_name)
-        cls.request.rows.row_ranges.add(**cls.row_range1.get_range_kwargs())
-        cls.request.rows.row_ranges.add(**cls.row_range2.get_range_kwargs())
-        cls.request.rows.row_ranges.add(**cls.row_range3.get_range_kwargs())
-
-    @staticmethod
-    def _get_target_class():
-        from google.cloud.bigtable.row_set import ReadRowsRequestManager
-        return ReadRowsRequestManager
-
-    def _make_one(self, *args, **kwargs):
-        return self._get_target_class()(*args, **kwargs)
-
-    def test_constructor(self):
-        request = mock.Mock()
-        last_scanned_key = "last_key"
-        rows_read_so_far = 10
-
-        request_manager = self._make_one(request, last_scanned_key,
-                                         rows_read_so_far)
-        self.assertEqual(request, request_manager.message)
-        self.assertEqual(last_scanned_key, request_manager.last_scanned_key)
-        self.assertEqual(rows_read_so_far, request_manager.rows_read_so_far)
-        self.assertEqual(None, request_manager.new_message)
-
-    def test__filter_row_key(self):
-        table_name = 'table_name'
-        request = _ReadRowsRequestPB(table_name=table_name)
-        request.rows.row_keys.extend([b'row_key1', b'row_key2',
-                                      b'row_key3', b'row_key4'])
-
-        last_scanned_key = b"row_key2"
-        request_manager = self._make_one(request, last_scanned_key, 2)
-        request_manager.new_message = _ReadRowsRequestPB(table_name=table_name)
-        request_manager._filter_rows_keys()
-
-        expected_request = _ReadRowsRequestPB(table_name=table_name)
-        expected_request.rows.row_keys.extend([b'row_key3', b'row_key4'])
-
-        self.assertEqual(expected_request, request_manager.new_message)
-
-    def test__filter_row_ranges_all_ranges_added_back(self):
-        last_scanned_key = _to_bytes("row_key14")
-        request_manager = self._make_one(self.request, last_scanned_key, 2)
-        request_manager.new_message = _ReadRowsRequestPB(
-            table_name=self.table_name)
-        request_manager._filter_row_ranges()
-
-        expected_request = _ReadRowsRequestPB(table_name=self.table_name)
-
-        expected_request.rows.row_ranges.add(**self.row_range1.
-                                             get_range_kwargs())
-        expected_request.rows.row_ranges.add(**self.row_range2.
-                                             get_range_kwargs())
-        expected_request.rows.row_ranges.add(**self.row_range3.
-                                             get_range_kwargs())
-
-        self.assertEqual(request_manager.new_message, expected_request)
-
-    def test__filter_row_ranges_all_ranges_already_read(self):
-        last_scanned_key = _to_bytes("row_key54")
-        request_manager = self._make_one(self.request, last_scanned_key, 2)
-        request_manager.new_message = _ReadRowsRequestPB(
-            table_name=self.table_name)
-        request_manager._filter_row_ranges()
-
-        expected_request = _ReadRowsRequestPB(table_name=self.table_name)
-        expected_request.rows.row_ranges.extend([])
-
-        self.assertEqual(request_manager.new_message, expected_request)
-
-    def test__filter_row_ranges_all_ranges_already_read_open_closed(self):
-        last_scanned_key = _to_bytes("row_key54")
-
-        row_range1 = RowRange("row_key21", "row_key29", False, True)
-        row_range2 = RowRange("row_key31", "row_key39")
-        row_range3 = RowRange("row_key41", "row_key49", False, True)
-
-        request = _ReadRowsRequestPB(table_name=self.table_name)
-        request.rows.row_ranges.add(**row_range1.get_range_kwargs())
-        request.rows.row_ranges.add(**row_range2.get_range_kwargs())
-        request.rows.row_ranges.add(**row_range3.get_range_kwargs())
-
-        request_manager = self._make_one(request, last_scanned_key, 2)
-        request_manager.new_message = _ReadRowsRequestPB(
-            table_name=self.table_name)
-        request_manager._filter_row_ranges()
-
-        expected_request = _ReadRowsRequestPB(table_name=self.table_name)
-        expected_request.rows.row_ranges.extend([])
-
-        self.assertEqual(request_manager.new_message, expected_request)
-
-    def test__filter_row_ranges_some_ranges_already_read(self):
-        last_scanned_key = _to_bytes("row_key22")
-        request_manager = self._make_one(self.request, last_scanned_key, 2)
-        request_manager.new_message = _ReadRowsRequestPB(
-            table_name=self.table_name)
-        request_manager._filter_row_ranges()
-
-        expected_request = _ReadRowsRequestPB(table_name=self.table_name)
-        row_range4 = RowRange("row_key22", "row_key29",
-                              start_inclusive=False)
-        expected_request.rows.row_ranges.add(**row_range4.
-                                             get_range_kwargs())
-        expected_request.rows.row_ranges.add(**self.row_range2.
-                                             get_range_kwargs())
-        expected_request.rows.row_ranges.add(**self.row_range3.
-                                             get_range_kwargs())
-
-        self.assertEqual(request_manager.new_message, expected_request)
-
-    def test_build_updated_request(self):
-        from google.cloud.bigtable.row_filters import RowSampleFilter
-        row_filter = RowSampleFilter(0.33)
-        last_scanned_key = _to_bytes("row_key14")
-        request = _ReadRowsRequestPB(filter=row_filter.to_pb(),
-                                     rows_limit=8,
-                                     table_name=self.table_name)
-        request.rows.row_ranges.add(**self.row_range1.get_range_kwargs())
-
-        request_manager = self._make_one(request, last_scanned_key, 2)
-
-        result = request_manager.build_updated_request()
-
-        expected_result = _ReadRowsRequestPB(table_name=self.table_name,
-                                             filter=row_filter.to_pb(),
-                                             rows_limit=6)
-        expected_result.rows.row_ranges.add(**self.row_range1.
-                                            get_range_kwargs())
-
-        self.assertEqual(expected_result, result)
-
-    def test__key_already_read(self):
-        last_scanned_key = _to_bytes("row_key14")
-        request = _ReadRowsRequestPB(table_name=self.table_name)
-        request_manager = self._make_one(request, last_scanned_key, 2)
-
-        self.assertTrue(request_manager._key_already_read(
-                                    _to_bytes("row_key11")))
-        self.assertFalse(request_manager._key_already_read(
-                                    _to_bytes("row_key16")))
 
 
 def _ReadRowsRequestPB(*args, **kw):
